geinos/requirements.txt
Flask
Flask-SqlAlchemy
mysql-connector-python
flask_httpauth
flask_restful
requests
lxml>=3.2.4
paramiko>=1.15.2
scp>=0.7.0
jinja2>=2.7.1
flask_cors
xmltodict
requests_ntlm
py-radius